pyproject.toml
[tool.poetry]
name = "fill_tables"
version = "0.1.0"
description = ""
authors = ["Xoma <xomyakes@gmail.com>"]
packages = [
  { include = "fill_tables" },
]

[tool.poetry.scripts]
fill_tables = "fill_tables.scripts.fill_tables:main"

[tool.poetry.dependencies]
python = "^3.9"
openpyxl = "^3.0.6"
httplib2 = "^0.19.0"
apiclient = "^1.0.4"
gspread = "^3.7.0"
google-auth = "^1.28.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
